chore(web-analytics): move the heatmaps scene into the product
The heatmaps backend already lives in products/web_analytics/backend, so
the frontend was the only half still sitting in frontend/src/scenes. The
scene registration moves from appScenes.ts, scenes.ts, and urls.ts into
the product manifest.

Lane effect: a heatmaps change claimed fe:core plus all 84 fe:product
lanes, because every file under frontend/ widens that way. It now claims
fe:product:web_analytics alone.

products/web_analytics/package.json gains the dependencies the moved code
imports. Files under products/ resolve from the product's own package
rather than from frontend/, so a dependency declared only in
frontend/package.json is unreachable after the move.
