+++ frontend/src/scenes/appScenes.ts
@@ -49,11 +49,7 @@ export const appScenes: Record<Scene | string, () => any> = {
     [Scene.Group]: () => import('./groups/Group'),
     [Scene.GroupsNew]: () => import('./groups/GroupsNew'),
     [Scene.Groups]: () => import('./groups/Groups'),
-    [Scene.Heatmaps]: () => import('./heatmaps/scenes/heatmaps/HeatmapsScene'),
     [Scene.Inbox]: () => import('./inbox/InboxScene'),
-    [Scene.HeatmapNew]: () => import('./heatmaps/scenes/heatmap/HeatmapNewScene'),
-    [Scene.HeatmapRecording]: () => import('./heatmaps/scenes/heatmap/HeatmapRecordingScene'),
-    [Scene.Heatmap]: () => import('./heatmaps/scenes/heatmap/HeatmapScene'),
     [Scene.HogFunction]: () => import('./hog-functions/HogFunctionScene'),
     [Scene.Insight]: () => import('./insights/InsightScene'),
     [Scene.InsightQuickStart]: () => import('./insights/InsightQuickStart/InsightQuickStart'),
