+++ frontend/src/scenes/scenes.ts
@@ -239,31 +239,10 @@ export const sceneConfigurations: Record<Scene | string, SceneConfig> = {
     },
     [Scene.GroupsNew]: { projectBased: true },
     [Scene.Groups]: { projectBased: true, name: 'Groups' },
-    [Scene.Heatmaps]: {
-        projectBased: true,
-        name: 'Heatmaps',
-        iconType: 'heatmap',
-        description: 'Heatmaps are a way to visualize user behavior on your website.',
-    },
     [Scene.Inbox]: {
         projectBased: true,
         name: 'Inbox',
         description: 'Actionable reports automatically generated from user session analysis and other signals.',
-    },
-    [Scene.Heatmap]: {
-        projectBased: true,
-        name: 'Heatmap',
-        iconType: 'heatmap',
-    },
-    [Scene.HeatmapNew]: {
-        projectBased: true,
-        name: 'New heatmap',
-        iconType: 'heatmap',
-    },
-    [Scene.HeatmapRecording]: {
-        projectBased: true,
-        name: 'Heatmap recording',
-        iconType: 'heatmap',
     },
     [Scene.HogFunction]: { projectBased: true, name: 'Hog function', activityScope: ActivityScope.HOG_FUNCTION },
     [Scene.Insight]: {
@@ -920,10 +899,6 @@ export const routes: Record<string, [Scene | string, string]> = {
     [urls.settings(':section' as any)]: [Scene.Settings, 'settings'],
     [urls.moveToPostHogCloud()]: [Scene.MoveToPostHogCloud, 'moveToPostHogCloud'],
     [urls.advancedActivityLogs()]: [Scene.AdvancedActivityLogs, 'advancedActivityLogs'],
-    [urls.heatmaps()]: [Scene.Heatmaps, 'heatmaps'],
-    [urls.heatmapNew()]: [Scene.HeatmapNew, 'heatmapNew'],
-    [urls.heatmapRecording()]: [Scene.HeatmapRecording, 'heatmapRecording'],
-    [urls.heatmap(':id')]: [Scene.Heatmap, 'heatmap'],
     [urls.liveDebugger()]: [Scene.LiveDebugger, 'liveDebugger'],
     [urls.links()]: [Scene.Links, 'links'],
     [urls.link(':id')]: [Scene.Link, 'link'],
