+++ frontend/src/scenes/session-recordings/player/sessionRecordingPlayerLogic.ts
@@ -45,7 +45,7 @@ import {
     MAX_REPLAY_IFRAME_HTML_CHARS,
     ReplayIframeData,
     persistReplayIframeData,
-} from 'scenes/heatmaps/replayIframeData'
+} from 'products/web_analytics/frontend/heatmaps/replayIframeData'
 import { preflightLogic } from 'scenes/PreflightCheck/preflightLogic'
 import { playerCommentModel } from 'scenes/session-recordings/player/commenting/playerCommentModel'
 import { sessionPlayerModalLogic } from 'scenes/session-recordings/player/modal/sessionPlayerModalLogic'
@@ -59,6 +59,7 @@ import { urls } from 'scenes/urls'
 import { userLogic } from 'scenes/userLogic'
 
 import { AvailableFeature, ExporterFormat, RecordingSegment, SessionPlayerData, SessionPlayerState } from '~/types'
+
 
 import type { FeatureFlagsSet } from '../../../lib/logic/featureFlagLogic'
 import type { PreflightStatus, SessionRecordingSnapshotSource, SessionRecordingType, UserType } from '../../../types'
