+++ products/web_analytics/frontend/heatmaps/components/ClickmapSettings.tsx
