+++ products/web_analytics/frontend/heatmaps/components/FilterPanel.tsx
