+++ products/web_analytics/frontend/heatmaps/components/FixedReplayHeatmapBrowser.tsx
@@ -2,8 +2,9 @@ import { useActions, useValues } from 'kea'
 import React from 'react'
 
 import { HeatmapCanvas } from 'lib/components/heatmaps/HeatmapCanvas'
-import { heatmapsBrowserLogic } from 'scenes/heatmaps/components/heatmapsBrowserLogic'
-import { RecordingClickmapOverlay } from 'scenes/heatmaps/components/RecordingClickmapOverlay'
+
+import { heatmapsBrowserLogic } from 'products/web_analytics/frontend/heatmaps/components/heatmapsBrowserLogic'
+import { RecordingClickmapOverlay } from 'products/web_analytics/frontend/heatmaps/components/RecordingClickmapOverlay'
 
 export function FixedReplayHeatmapBrowser({
     iframeRef,
