+++ products/web_analytics/frontend/heatmaps/components/HeatmapRecordingFallback.tsx
