+++ products/web_analytics/frontend/heatmaps/components/HeatmapsEnableCapture.tsx
