+++ products/web_analytics/frontend/heatmaps/components/HeatmapsWarnings.tsx
