+++ products/web_analytics/frontend/heatmaps/components/RecordingClickmapOverlay.tsx
