+++ products/web_analytics/frontend/heatmaps/components/heatmapRecordingFallbackLogic.ts
