+++ products/web_analytics/frontend/heatmaps/components/heatmapsBrowserLogic.test.ts
