+++ products/web_analytics/frontend/heatmaps/components/heatmapsBrowserLogic.ts
@@ -22,8 +22,11 @@ import {
     authorizedUrlListLogic,
     defaultAuthorizedUrlProperties,
 } from 'lib/components/AuthorizedUrlList/authorizedUrlListLogic'
+import type { KeyedAppUrl } from 'lib/components/AuthorizedUrlList/authorizedUrlListLogic'
 import { heatmapDataLogic } from 'lib/components/heatmaps/heatmapDataLogic'
+import type { HrefMatchType } from 'lib/components/heatmaps/heatmapDataLogic'
 import { CommonFilters, HeatmapFixedPositionMode } from 'lib/components/heatmaps/types'
+import type { HeatmapFilters } from 'lib/components/heatmaps/types'
 import { PostHogAppToolbarEvent, calculateViewportRange } from 'lib/components/IframedToolbarBrowser/utils'
 import { FEATURE_FLAGS } from 'lib/constants'
 import { LemonBannerProps } from 'lib/lemon-ui/LemonBanner'
@@ -40,10 +43,6 @@ import { hogql } from '~/queries/utils'
 
 import { savedPreflightCreate } from 'products/web_analytics/frontend/generated/api'
 import type { HeatmapPreflightResponseApi } from 'products/web_analytics/frontend/generated/api.schemas'
-
-import type { KeyedAppUrl } from '../../../lib/components/AuthorizedUrlList/authorizedUrlListLogic'
-import type { HrefMatchType } from '../../../lib/components/heatmaps/heatmapDataLogic'
-import type { HeatmapFilters } from '../../../lib/components/heatmaps/types'
 
 export type HeatmapsBrowserLogicProps = {
     iframeRef: RefObject<HTMLIFrameElement | null>
