+++ products/web_analytics/frontend/heatmaps/components/recordingClickmapLogic.test.ts
