+++ products/web_analytics/frontend/heatmaps/replayIframeData.test.ts
