+++ products/web_analytics/frontend/heatmaps/replayIframeData.ts
