+++ products/web_analytics/frontend/heatmaps/scenes/heatmap/HeatmapNewScene.stories.tsx
