+++ products/web_analytics/frontend/heatmaps/scenes/heatmap/HeatmapScene.stories.tsx
