+++ products/web_analytics/frontend/heatmaps/scenes/heatmap/heatmapCreationLogic.test.ts
