+++ products/web_analytics/frontend/heatmaps/scenes/heatmap/heatmapCreationLogic.ts
@@ -24,19 +24,19 @@ import { FEATURE_FLAGS } from 'lib/constants'
 import { FeatureFlagsSet, featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { getAccessControlDisabledReason } from 'lib/utils/accessControlUtils'
 import { inStorybook, inStorybookTestRunner } from 'lib/utils/dom'
-import { heatmapsBrowserLogic, isUrlPattern } from 'scenes/heatmaps/components/heatmapsBrowserLogic'
-import { ReplayIframeData, getStoredRecordingBackground } from 'scenes/heatmaps/replayIframeData'
 import { sessionPlayerModalLogic } from 'scenes/session-recordings/player/modal/sessionPlayerModalLogic'
+import type { SessionPlayerModalContext } from 'scenes/session-recordings/player/modal/sessionPlayerModalLogic'
 import { teamLogic } from 'scenes/teamLogic'
 import { urls } from 'scenes/urls'
 
 import { hogql } from '~/queries/utils'
 import { AccessControlLevel, AccessControlResourceType, HeatmapType } from '~/types'
+import type { TeamPublicType, TeamType } from '~/types'
 
 import { savedPrewarmCreate } from 'products/web_analytics/frontend/generated/api'
+import { heatmapsBrowserLogic, isUrlPattern } from '../../components/heatmapsBrowserLogic'
+import { ReplayIframeData, getStoredRecordingBackground } from '../../replayIframeData'
 
-import type { TeamPublicType, TeamType } from '../../../../types'
-import type { SessionPlayerModalContext } from '../../../session-recordings/player/modal/sessionPlayerModalLogic'
 import type { PagePreflight } from '../../components/heatmapsBrowserLogic'
 import { HeatmapCreationContext, heatmapLogic } from './heatmapLogic'
