+++ products/web_analytics/frontend/heatmaps/scenes/heatmap/heatmapLogic.test.ts
