+++ products/web_analytics/frontend/heatmaps/scenes/heatmap/heatmapRecordingLogic.ts
