+++ products/web_analytics/frontend/heatmaps/scenes/heatmaps/HeatmapsScene.stories.tsx
