+++ products/web_analytics/frontend/heatmaps/scenes/heatmaps/heatmapsSceneLogic.ts
